requirements.txt
Flask==3.0.2
Flask-API==3.1
Flask-RESTful==0.3.10
graphene==3.3
pymongo==4.6.3
requests~=2.31.0
pymongo[srv]
python-dotenv~=0.21.0